cafekiosk/src/main/java/sample/cafekiosk/spring/api/service/product/ProductNumberFactory.java
package sample.cafekiosk.spring.api.service.product;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;
import sample.cafekiosk.spring.domain.product.ProductRepository;

@Component
@RequiredArgsConstructor
public class ProductNumberFactory {

    private final ProductRepository productRepository;

    public String createNextProductNumber() {
        // productNumber
        // 001 002 003 004
        // DB에서 마지막 저장된 Product의 상품번호를 읽어와서 + 1
        String latestProductNumber = productRepository.findLatestProductNumber();
        if (latestProductNumber == null) {
            return "001";
        }
//        int nextProductNumber = Integer.parseInt(latestProductNumber) + 1;
        int latestProductNumberInt = Integer.parseInt(latestProductNumber);
        int nextProductNumberInt = latestProductNumberInt + 1;

        // 9 -> 009, 10 -> 010
        return String.format("%03d", nextProductNumberInt);

    }
}
